backend/pyproject.toml
[project]
name = "sentinel-backend"
version = "0.3.0"
description = "Sentinel backend and incident intelligence agents"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "mangum>=0.19.0",
  "pydantic>=2.7.0",
  "boto3>=1.34.0",
  "uvicorn>=0.30.0",
  "httpx>=0.27.0",
  "PyJWT[crypto]>=2.8.0",
  "fpdf2>=2.8.7",
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
]
